requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg[binary]
alembic
pydantic
pydantic-settings
python-dotenv
yfinance
requests
aiohttp
sentence-transformers
chromadb
google-generativeai
langchain-core
langchain-google-genai
pytest
pytest-asyncio
httpx
greenlet
tenacity
langchain-groq
